face_recognition_service/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
numpy==2.3.1
opencv-python==4.11.0.86
Pillow==11.2.1
requests==2.32.3